Move internal/log to pkg/log
