+++ internal/rpc/http.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/NethermindEth/juno/internal/log"
+	"github.com/NethermindEth/juno/pkg/log"
 	"github.com/NethermindEth/juno/pkg/jsonrpc"
 	. "github.com/NethermindEth/juno/pkg/jsonrpc/providers/http"
 )
